LayoutTests/media/modern-media-controls/tracks-support/tracks-support-show-panel-then-double-click-on-tracks-button-expected.txt

Ensure that double-clicking the tracks button when the tracks panel makes the panel appear again.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media started playing
PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
PASS shadowRoot.querySelector('.tracks-panel') became different from null
PASS shadowRoot.querySelector('.tracks-panel') is not null

PASS successfullyParsed is true

TEST COMPLETE
